SingleElementLayout -> UIBox; console command tweaks
