+++ engine/src/main/resources/assets/skins/nuiEditorScreen.skin
@@ -1,3 +1,10 @@
 {
-    "inherit": "engine:mainMenu"
+    "inherit": "engine:mainMenu",
+    "elements": {
+        "UIBox": {
+            "background": "",
+            "background-border": {},
+            "margin": {}
+        }
+    }
 }
