+++ engine/src/main/resources/assets/ui/nuiEditorScreen.ui
@@ -9,7 +9,7 @@
         "debug": true,
         "contents": [
             {
-                "type": "SingleElementLayout",
+                "type": "UIBox",
                 "id": "selectedScreen",
                 "updateContent": false,
                 "layoutInfo": {
